+++ certbot/main.py
@@ -1368,6 +1368,10 @@ def main(cli_args=None):
         if config.func != plugins_cmd:
             raise
 
+    if sys.version_info[:2] == (3, 4):
+        logger.warning("Python 3.4 support will be dropped in the next release "
+                    "of Certbot - please upgrade your Python version.")
+
     set_displayer(config)
 
     # Reporter